requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
benzinga==1.21
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.0
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
h11==0.14.0
huggingface-hub==0.16.2
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
jsonschema==4.17.3
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
numpy==1.25.0
openpyn==3.0.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.2
pathy==0.10.2
preshed==3.0.8
prettytable==3.8.0
pycparser==2.21
pydantic==1.10.10
pyreadline3==3.4.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
selenium==4.10.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
structlog==23.1.0
sympy==1.12
termcolor==2.3.0
thinc==8.1.10
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
trio==0.22.0
trio-websocket==0.10.3
typer==0.9.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==1.25.10
verboselogs==1.7
wasabi==1.1.2
wcwidth==0.2.6
wsproto==1.2.0
yarl==1.9.2
youtube-transcript-api==0.6.1
